==> notebooks/02_Weather_Data_Cleaning.ipynb <==
# Cleaning the weather dataset

## Importing data and modules

import pandas as pd

weather = pd.read_csv('../data/weather.csv')

## Checking for null values, data shape

weather.columns

weather.shape

weather.isnull().sum().sum()

weather.info()

The only non object features are `Station`, `Tmax`, `Tmin`, `Dewpoint`, `ResultSpeed`, and `ResultDir`. The built in pandas methods for checking for missing data should suffice to sort any missing values for them. For the remaining features we have to look more closely. 

weather.Station.value_counts()

weather.Tmax.isnull().sum()

weather.Tmin.isnull().sum()

weather.DewPoint.isnull().sum()

weather.ResultSpeed.isnull().sum()

weather.ResultDir.isnull().sum()

## Checking for missing data

**So there are no true null values in the columns that are already coded as numeric, but we might need to look more carefully to ensure that missing values were not encoded as strings or some other data type. According to the data dictionary, most features encode missing values with an 'M'.**

### Date

The date column is in a consistent format (yyyy-mm-dd) that we can use for merging:

weather.Date.describe()

weather.Date.map(len).value_counts()

### Tavg

There are very few missing values for average temperature; let's drop those rows.

Tavg_missing = weather['Tavg'].map(lambda val: val == 'M')
Tavg_missing.sum()

weather = weather[~Tavg_missing]

weather['Tavg'] = weather['Tavg'].astype(float)

Depart_missing = weather['Depart'].map(lambda val: val == 'M')
Depart_missing.sum()

weather.drop('Depart', 1, inplace=True)

### WetBulb

Again, we have very few missing values for the wetbulb temperature. We can drop those entries as well.

WetBulb_missing = weather['WetBulb'].map(lambda val: val == 'M')
WetBulb_missing.sum()

weather = weather[~WetBulb_missing]

weather['WetBulb'] = weather['WetBulb'].astype(float)

### Heat, Cool

We can convert these variables to numeric after ensuring that no 'M's remain:

Heat_missing = weather['Heat'].map(lambda val: val == 'M')
Heat_missing.sum()

weather['Heat'] = weather['Heat'].astype(float)

Cool_missing = weather['Cool'].map(lambda val: val == 'M')
Cool_missing.sum()

weather['Cool'] = weather['Cool'].astype(float)

### Sunrise, Sunset

Sunrise_missing = weather['Sunrise'].map(lambda val: val == 'M')
Sunrise_missing.sum()

weather['Sunrise'].value_counts().head()

Sunset_missing = weather['Sunset'].map(lambda val: val == 'M')
Sunset_missing.sum()

weather['Sunset'].value_counts().head()

### CodeSum

Here, we split the codesum column into separate binary columns for each individual code:

CodeSum_missing = weather['CodeSum'].map(lambda val: val == 'M')
CodeSum_missing.sum()

codesums  = set(weather.CodeSum.str.split(' ').map(tuple))

codes = set()
for cs in codesums:
    for code in cs:
        codes.add(code)
codes.discard('')
codes = list(codes)

def build_df_codesum(row, codes):
    cs = set(row['CodeSum'].split())
    code_out = [int(code in cs) for code in codes]
    code_out = pd.Series(data=code_out, index=codes)
    return code_out

weather[codes] = weather.apply(lambda row:build_df_codesum(row, codes), 1)

weather.drop('CodeSum', 1, inplace=True)

### Depth

Since around half of all depth values were missing, we chose to drop this column:

Depth_missing = weather['Depth'].map(lambda val: val == 'M')
Depth_missing.sum()

weather['Depth'].value_counts()

weather.drop('Depth', 1, inplace=True)

### Water1, SnowFall, PrecipTotal, 

Water1_missing = weather['Water1'].map(lambda val: val == 'M')
Water1_missing.sum()

weather['Water1'].value_counts()

weather.drop('Water1', 1, inplace=True)

It's very unlikely that snow would be falling during the summer months when the weather data was collected; we can see that there were only 13 instances of any evidence of snowfall. Let's drop the snowfall column.

SnowFall_missing = weather['SnowFall'].map(lambda val: val == 'M')
SnowFall_missing.sum()

weather['SnowFall'].value_counts()

weather.drop('SnowFall', 1, inplace=True)

PrecipTotal_missing = weather['PrecipTotal'].map(lambda val: val == 'M')
PrecipTotal_missing.sum()

weather = weather[~PrecipTotal_missing]

weather.PrecipTotal = weather.PrecipTotal.map(lambda x: x.replace('T', '0.001')).astype(float)

## StnPressure

StnPressure_missing = weather['StnPressure'].map(lambda val: val == 'M')
StnPressure_missing.sum()

weather = weather[~StnPressure_missing]  
weather['StnPressure'] = weather['StnPressure'].map(float)

## SeaLevel

SeaLevel_missing = weather['SeaLevel'].map(lambda val: val == 'M')
SeaLevel_missing.sum()

weather = weather[~SeaLevel_missing]  
weather['SeaLevel'] = weather['SeaLevel'].map(float)

## AvgSpeed

AvgSpeed_missing = weather['AvgSpeed'].map(lambda val: val == 'M')
AvgSpeed_missing.sum()

weather['AvgSpeed'] = weather['AvgSpeed'].map(float)

## Exporting to csv

weather.head()

weather.isnull().sum().sum()

weather.to_csv('../data/weather_cleaned.csv', index=False)
