==> notebooks/03_Spray_Data_Cleaning.ipynb <==
# Cleaning the Spray Dataset

## Importing modules and data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import datetime as dt

%matplotlib inline
sns.set_style('darkgrid')
sns.set_palette('viridis')

spray = pd.read_csv('../data/spray.csv')

spray.head()

## Checking for null values

spray.isnull().sum()

Since there are so many missing values for time and we have no way to determine the true values, we will drop the time column.

## Checking for other errors

### Duplicated locations

We noticed that one latitude/longitude pair was duplicated over 500 times and one other was duplicated once. We address these issues below:

order_groups = spray.groupby(['Latitude', 'Longitude'], as_index=False).count().sort_values('Date', ascending=False)

order_groups.head()

order_groups.loc[11499, :]['Latitude']

lat_long_count = spray.drop('Time', axis=1).groupby(['Latitude', 'Longitude']).agg(['count'])

lat_long_count.columns

lat_long_count = pd.Series(lat_long_count.values.reshape(1,-1)[0], lat_long_count.index)

lat_long_count.sort_values(ascending=False).head()

Through manual inspection, we have determined a spray entry which has been duplicated 541 times.  We are going to remove the duplicates from this list.

mask = (spray['Latitude'] == 41.986460) & (spray['Longitude'] == -87.794225)
spray[mask].shape

spray[mask].index[1:]

spray.drop(spray[mask].index[1:], inplace=True)

lat_long_count = spray.drop('Time', axis=1).groupby(['Latitude', 'Longitude']).agg(['count'])

lat_long_count = pd.Series(lat_long_count.values.reshape(1,-1)[0], lat_long_count.index)

lat_long_count.sort_values(ascending=False).head()

mask = (spray['Longitude'] == -87.7930883333333) & (spray['Latitude'] ==  41.9839166666667)
sum(mask)

spray[mask].index[1:]

spray.drop(spray[mask].index[1:], inplace=True)

lat_long_count = spray.drop('Time', axis=1).groupby(['Latitude', 'Longitude']).agg(['count'])

lat_long_count = pd.Series(lat_long_count.values.reshape(1,-1)[0], lat_long_count.index)

lat_long_count.sort_values(ascending=False).head()

spray.drop('Time', axis=1, inplace=True)

**Now there are no duplicated latitude/longitude pairs and we have dropped the time column.**

## Adding an indicator if a trap is nearby spray zone

train = pd.read_csv('../data/train_cleaned.csv')

from math import sin, cos, radians, asin, sqrt
def global_distance(lon1, lat1, lon2, lat2):
    # convert decimal degrees to radians 
    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])

    dlon = lon2 - lon1 
    dlat = lat2 - lat1 
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * asin(sqrt(a)) 
    r = 3956
    
    return c * r

train[['longitude','latitude']].head()

spray[['Longitude', 'Latitude']].head()

stations = {}
for index, row in train.iterrows():
    stations[row['trap']] = (row['longitude'], row['latitude'])

stations

station_distances = {}
for station in stations:
    lon = stations[station][0]
    lat = stations[station][1]
    
    for index, spray_row in spray.iterrows():
        tmp_dist = global_distance(lon, lat, spray_row['Longitude'], spray_row['Latitude'])
        if station in station_distances:
            station_distances[station] = min(tmp_dist, station_distances[station])
        else:
            station_distances[station] = tmp_dist


train.to_csv('../data/train_cleaned_spray.csv', index=False)

## Exporting file

spray.to_csv('../data/spray_cleaned.csv', index=False)
