==> notebooks/04_Data_Merging.ipynb <==
# Data Merging

## Importing data and modules

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline
sns.set_style('darkgrid')
sns.set_palette('viridis')

train = pd.read_csv('../data/train_cleaned_spray.csv')
weather = pd.read_csv('../data/weather_cleaned.csv')
spray = pd.read_csv('../data/spray_cleaned.csv')

weather.columns = weather.columns.map(lambda x: x.lower())
spray.columns = spray.columns.map(lambda x: x.lower())

train.date[0], weather.date[0], spray.date[0]

The date formats in each of the datasets are the same so we can merge on those columns.

## Establishing `station` column in train dataset to merge

We decide to split Chicago on a E/W axis to divide the city into north and south halves. Traps in the 'north' half will be tied to weather recorded at O'Hare Airport (`station` = 1) and traps in the 'south' half will be tied to weather recorded at Midway Airport (`station` = 2). To do this, we find the midpoint between the latitudes of both stations and use that latitude line as the division between the north and south sides of Chicago.

### Finding the midpoint

(41.998+41.786)/2

### Using the midpoint to divide observations

train['station'] = np.where(train['latitude']>=41.892, 1, 2)

train.station.value_counts()

train.shape, weather.shape

train_weather = pd.merge(train, weather, on=['date', 'station'], )

train_weather.shape

train_weather.columns

## Exporting merged data

train_weather.to_csv('../data/train_weather_spray_merged.csv', index=False)

train_weather.isnull().any()
