==> notebooks/06_Modeling_JB.ipynb <==
# Modeling

## Importing modules and data

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier, BaggingClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score, roc_auc_score
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import pandas as pd
import numpy as np
import random
import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline
sns.set_style('darkgrid')

np.random.seed(32)

df = pd.read_csv('../data/train_weather_spray_merged.csv')

df_dummied = pd.get_dummies(df, columns=['species'])

## Creating Validation Set, Scaling

df_dummied.drop(columns=['date', 'address', 'block', 'street', 'trap', 'addressnumberandstreet', 'nummosquitos', 'sunrise', 'sunset'], axis=1, inplace=True)

features = [col for col in df_dummied if col != 'wnvpresent']

df_dummied.columns

X = df_dummied[features]
y = df_dummied.wnvpresent

X.columns

X_train, X_test, y_train, y_test = train_test_split(X, y)

ss = StandardScaler()
X_train = ss.fit_transform(X_train)
X_test = ss.transform(X_test)

## Logistic Regression

logreg = LogisticRegression()
gs_logreg = GridSearchCV(logreg, param_grid = {'penalty': ['l1', 'l2'],
                                               'C': np.linspace(1, 10, 10),
                                               'class_weight':['balanced']}, scoring='roc_auc', verbose=1)

%%time
gs_logreg.fit(X_train, y_train)

gs_logreg.best_params_

gs_logreg.score(X_train, y_train), gs_logreg.score(X_test, y_test)

## Decision Tree

gs_dt = GridSearchCV(DecisionTreeClassifier(), param_grid={'min_samples_split' : [2, 3, 4, 5],
                                                           'min_samples_leaf': [2]},
                                               scoring='roc_auc')

gs_dt.fit(X_train, y_train)

gs_dt.score(X_train, y_train), gs_dt.score(X_test, y_test)

## Random Forest

gs_rf = GridSearchCV(RandomForestClassifier(), param_grid={'min_samples_split': [2,3,4],
                                                           'min_samples_leaf': [2],
                                                           'class_weight': ['balanced'],},
                                                           scoring='roc_auc')

gs_rf.fit(X_train, y_train)

gs_rf.score(X_train, y_train), gs_rf.score(X_test, y_test)

gs_rf.best_params_

#### Support Vector Classifier

param_grid = {
    'kernel': ['rbf', 'sigmoid'],
    'gamma': np.logspace(-5, 2, 20)
}

clf = SVC(probability=True)

grid = GridSearchCV(clf, param_grid, cv=5, scoring='roc_auc', verbose=1)
grid.fit(X_train, y_train)

print(grid.best_params_)
print(grid.best_score_)

grid.score(X_test, y_test)

gs_rf = GridSearchCV(RandomForestClassifier(), param_grid={'n_estimators': range(10, 51, 10),
                                                          'max_depth': [7, 15, 25, 35],
                                                          'min_samples_leaf': [2, 4, 8],},
                                                           scoring='roc_auc')

gs_rf.fit(X_train, y_train)

print(gs_rf.best_params_)
print(gs_rf.best_score_)

import pickle

with open('../data/rf_01.pkl', 'wb+') as f:
    pickle.dump(gs_rf, f)

## New Params

from sklearn.metrics import roc_auc_score

rf_2 = RandomForestClassifier(n_estimators=1000, max_depth=7, min_samples_leaf=4)

rf_2.fit(X_train, y_train)

roc_auc_score(y_train, rf_2.predict_proba(X_train)[:,1])

roc_auc_score(y_test, rf_2.predict_proba(X_test)[:, 1])

rf_2.score(X_train, y_train)

rf_2.score(X_test, y_test)

gs_rf = GridSearchCV(RandomForestClassifier(), param_grid={'n_estimators': [1000],
                                                          'max_depth': [5, 7, 15],
                                                          'min_samples_leaf': [2, 4, 8],},
                                                           scoring='roc_auc')

gs_rf.fit(X_train, y_train)

print(gs_rf.best_params_)
print(gs_rf.best_score_)

test = pd.read_csv('../data/test_merged.csv')

test = ss.fit_transform(test)

preds = grid.predict_proba(test)[:, 1]

preds

preds_forest = gs_rf.predict_proba(test)[:, 1]

submission = pd.DataFrame(preds, index=None, columns=['WnvPresent'])

index_array = np.array(range(1,116294))

submission = submission.set_index(index_array)

submission.index.rename('Id', inplace=True)

submission.to_csv('../data/Submission_1')

submission_forest = pd.DataFrame(preds_forest, index=None, columns=['WnvPresent'])

index_array = np.array(range(1,116294))

submission_forest = submission_forest.set_index(index_array)

submission_forest.index.rename('Id', inplace=True)


submission_forest.to_csv('../data/Submission_2')

preds_forest = gs_rf.predict_proba(test)[:,1]

logreg = LogisticRegression()
gs_logreg = GridSearchCV(logreg, param_grid = {'penalty': ['l1', 'l2'],
                                               'C': np.linspace(1, 10, 10),
                                               'class_weight':['balanced']}, scoring='roc_auc', verbose=1)

gs_logreg.fit(X_train, y_train)

print(gs_logreg.best_params_)
print(gs_logreg.best_score_)

preds_log = gs_logreg.predict_proba(test)[:,1]

submission_log = pd.DataFrame(preds_log, index=None, columns=['WnvPresent'])

index_array = np.array(range(1,116294))

submission_log = submission_log.set_index(index_array)

submission_log.index.rename('Id', inplace=True)


submission_log.to_csv('../data/Submission_3')

rf_2_preds = rf_2.predict_proba(test)[:, 1]

submission_rf_2 = pd.DataFrame(rf_2_preds, index=None, columns=['WnvPresent'])

index_array = np.array(range(1,116294))

submission_rf_2 = submission_rf_2.set_index(index_array)

submission_rf_2.index.rename('Id', inplace=True)


submission_rf_2.to_csv('../data/Submission_4')

End of Notebook
